Split Activities class into isolated Activity classes
